+++ src/main/java/com/open3hr/adeies/controllers/UserController.java
@@ -16,7 +16,7 @@ public class UserController {
 
         @GetMapping("/{id}" )
         @CrossOrigin(origins = "*", allowedHeaders = "*", methods = {RequestMethod.GET, RequestMethod.POST, RequestMethod.PUT, RequestMethod.DELETE})
-        public UserDTO getUserById(@PathVariable Long id){
+        public UserDTO getUserById(@PathVariable Integer id){
             return userService.findById(id);
         }
 
@@ -38,7 +38,7 @@ public class UserController {
         }
 
         @DeleteMapping("/{id}")
-        public void deleteById(@PathVariable Long id){
+        public void deleteById(@PathVariable Integer id){
                 userService.deleteById(id);
         }
 }
