+++ src/main/java/com/open3hr/adeies/dto/LeaveBalanceDTO.java
@@ -11,11 +11,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class LeaveBalanceDTO {
-    private Long id;
-    private Long days;
-    private Long daysTaken;
-    private Long employeeId;
-    private Long leaveCategoryId;
+    private Integer id;
+    private Integer days;
+    private Integer daysTaken;
+    private Integer employeeId;
+    private Integer leaveCategoryId;
 
     public LeaveBalanceDTO(LeaveBalance leaveBalance){
         this.id = leaveBalance.getId();
